% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testCompRiskReg.R
\name{compRiskReg}
\alias{compRiskReg}
\alias{fitReg}
\title{making cuminc data for plotting and testing}
\usage{
fitReg(time, risk, group, data, cens = 0)
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{conf.int}{level of two sided conf int}
}
\description{
cox models for competing risks and differences in them between groups
}
