% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsTab.R
\name{eventsTab}
\alias{eventsTab}
\alias{eventTab}
\title{number of events up to time t}
\usage{
eventTab(fit, ci, risk, group)
}
\arguments{
\item{fit}{fitSurvival object}
}
\description{
tables with number of events up to time t, for given risk and group
}
