% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testCuminc.R
\name{fitCumul}
\alias{fitCumul}
\alias{fitCuminc}
\title{fitting cuminc object}
\usage{
fitCuminc(time, risk, group, data, cens = 0, rho = 0)
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}
}
\description{
list with cumulative incidences and a test of differences between them
}
