% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testSurvival.R
\name{fitSurvival}
\alias{fitSurvival}
\title{simple survfit for each risk separetly}
\usage{
fitSurvival(time, risk, group, data, cens = 0, type = "kaplan-meier",
  conf.int = 0.95, conf.type = "log")
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can
be data frame or matrix}

\item{type}{"kaplan-meier", "fleming-harrington" or "fh2"}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}

\item{error}{"greenwood (default)', "tsiatis", "aalen}
}
\description{
The function fits Kaplan-Mayer models for each risk
treating other events as censoring
}
