% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crSummary.R
\name{fullAnalysis}
\alias{fullAnalysis}
\alias{crSummary}
\title{full analysis of data}
\usage{
crSummary(time, risk, group, data, target, cens = 0, rho = 0,
  type = "kaplan-meier", conf.int = 0.95, conf.type = "log")
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can
be data frame or matrix}

\item{type}{"kaplan-meier", "fleming-harrington" or "fh2"}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}

\item{error}{"greenwood (default)', "tsiatis", "aalen}
}
\description{
The functionn
}
