% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestSurvival.R
\name{lrtSurvival}
\alias{lrtSurvival}
\title{Fleming-Harrington test for differences between groups.}
\usage{
lrtSurvival(time, risk, group, data, cens = 0, rho = 0)
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{nam of a column indicating group variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containg time, risk and group columns.}

\item{cens}{value of 'risk' indicating censored observation.}

\item{rho}{rho parameter from Fleming-Harrington Test.}
}
\value{
a data.frame containing p-values of Fleming-Harrington Test for each risk.
}
\description{
The function tests, if there are differences between groups for survival curves estimating for all risks separately (treating other events as censoring).
}
\examples{
lrtSurvival(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive", rho = 0)
}
