% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUAD.R
\docType{data}
\name{LUAD}
\alias{LUAD}
\title{LUAD dataset from The Cancer Genome Atlas Project.}
\description{
LUAD dataset contains following information about patients with lung cancer:

- "time" is a time from beginning of an observation to an event or follow-up,

- "event" is a type of event that have happened ("death", "new_tumor" or "alive" when the observation was censored),

- "gender" is a gender of a patient (grouping variable).

The dataset is published in ‘RTCGA.clinical’ package: https://bioconductor.org/packages/release/data/experiment/manuals/RTCGA.clinical/man/RTCGA.clinical.pdf.
}
