% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvival.R
\name{plotSurvival}
\alias{plotSurvival}
\title{ploting}
\usage{
plotSurvival(fit, target)
}
\arguments{
\item{time}{time must be numeric}

\item{status}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{type}{"kaplan-meier", "fleming-harrington" or "fh2"}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}

\item{error}{"greenwood (default)', "tsiatis", "aalen}
}
\description{
The function plots the survival curves for both risks and within all groups
}
