% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testCompRiskReg.R
\name{regTest}
\alias{regTest}
\title{making cuminc data for plotting and testing}
\usage{
regTest(reg, conf.int = 0.95)
}
\arguments{
\item{reg}{time must be numeric}

\item{conf.int}{level of two sided conf int}
}
\description{
cox models for competing risks and differences in them between groups
}
