% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskTab.R
\name{riskTab}
\alias{riskTab}
\title{Number at risk table.}
\usage{
riskTab(time, risk, group, data, cens = 0)
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{name of a column indicating group variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containing time, risk and group columns.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}
}
\value{
A grob with n tables, where n is number of risks. Each table contains number of observations at risk in each group in given time points (the time points correspond to breaks at x-axis of plots with survival curves).
}
\description{
The function creates tables for each risk with number of observation at risk in given time in groups.
}
\examples{
riskTab(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
}
