% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testCox.R
\name{testCox}
\alias{testCox}
\title{test cox model}
\usage{
testCox(fitCox)
}
\arguments{
\item{fitCox}{time must be numeric}
}
\description{
The function fits cos models for each risk
treating other events as censoring
}
