% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCompRiskReg.R
\name{compRiskReg}
\alias{compRiskReg}
\alias{fitReg}
\title{Regression Models for Competing Risks.}
\usage{
fitReg(time, risk, group, cens = 0)
}
\arguments{
\item{time}{vector with times of the first event or follow-up, must be numeric.}

\item{risk}{vector with type of event, can be numeric or factor/character.}

\item{group}{vector with group variable, can be numeric or factor/character.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}
}
\value{
a list of length n, where n is number of different types of events. Each element of a list contains a result of crr function from cmprsk package for given type of event.
}
\description{
Fitting Cox model (Regression model) for competing risks.

The function fits survival curves for each risk treating other events as censoring.
}
\examples{
fitReg(time = LUAD$time, risk = LUAD$event, group = LUAD$gender, cens = "alive")
}
