% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCompRiskReg.R
\name{compRiskReg}
\alias{compRiskReg}
\alias{fitReg}
\title{Regression Models for Competing Risks.}
\usage{
fitReg(time, risk, group, data, cens = 0)
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{nam of a column indicating group variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containg time, risk and group columns.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}
}
\value{
a list of length n, where n is number of different types of events. Each element of a list contains a result of crr function from cmprsk package for given type of event.
}
\description{
Fitting Cox model (Regression model) for competing risks.
}
\examples{
fitReg(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
}
