% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crSummary.R
\name{crSummary}
\alias{crSummary}
\title{Competing Risks Models Summary.}
\usage{
crSummary(time, risk, group, data, target, cens = 0, rho = 0,
  type = "kaplan-meier", conf.int = 0.95, conf.type = "log")
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{nam of a column indicating grouping variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containing time, risk and group columns.}

\item{target}{point in time, in which the confidence bounds should be plotted.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}

\item{type}{type of survival curve to be fitted. Possible values are "kaplan-meier" (default), "fleming-harrington" or "fh2".}

\item{conf.int}{level of two sided confidence interval.}

\item{conf.type}{type of confidence interval. Possilble values: "none", "plain", "log" (default), "log-log".}
}
\value{
Results of functions implemented in the package summarised in an one-page raport.
}
\description{
The function generates summariezed raport including
p-values of testing differences between groups and visualisation of survival
and cumulative incidences curves.
}
\examples{
crSummary(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive", target = 1200, type = "kaplan-meier", conf.int = 0.95, conf.type = "log")
}
