% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsTab.R
\name{eventsTab}
\alias{eventsTab}
\alias{eventTab}
\title{Number of events tables.}
\usage{
eventTab(fit, ci, risk, group)
}
\arguments{
\item{fit}{a result of fitSurvival function.}

\item{ci}{a result of fitCuminc function.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{name of a column indicating group variable, can be numeric or factor/character.}
}
\value{
A grob with n tables, where n is number of risks. Each table contains number of events that have happened in each group up to given time point (the time points correspond to breaks at x-axis of plots with cumulative incidence curves).
}
\description{
The function creates tables for each risk with number of events up to given time in groups.
}
\examples{
fitS <- fitSurvival(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive", type = "kaplan-meier", conf.int = 0.95, conf.type = "log")
fitC <- fitCuminc(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
eventTab(fit = fitS, ci = fitC, risk = "event", group = "gender")
}
