% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCuminc.R
\name{fitCumul}
\alias{fitCumul}
\alias{fitCuminc}
\title{Cumulative Incidences Curves.}
\usage{
fitCuminc(time, risk, group, data, cens = 0)
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{nam of a column indicating group variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containg time, risk and group columns.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}
}
\value{
list of length [(number of risks)*(number of groups) + 1], containing estimation of cumulative incidences curves for each risk and group. The last element of a group is a data.frame with results of a K-sample test.
}
\description{
Fitting cumulative incidence function across different groups and risks.
}
\examples{
fitCuminc(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
}
