% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtSurvival.R
\name{lrtSurvival}
\alias{lrtSurvival}
\title{Log rank test for differences between groups}
\usage{
lrtSurvival(time, risk, group, data, cens = 0, rho = 0)
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{type}{"kaplan-meier", "fleming-harrington" or "fh2"}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}

\item{error}{"greenwood (default)', "tsiatis", "aalen}
}
\description{
The function tests, if there are differences between groups for two risks independently
}
