% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCuminc.R
\name{plotCuminc}
\alias{plotCuminc}
\title{Cumulative incidences curves}
\usage{
plotCuminc(ci, risk, group, target)
}
\arguments{
\item{ci}{a result of function fitCuminc.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{name of a column indicating group variable, can be numeric or factor/character.}

\item{target}{point in time, in which the confidence bounds should be plotted.}
}
\value{
a ggplot containing n graphs, where n is number of risks. Each graphs represents cumulative incidence curves for given risk in each group.
}
\description{
The function plots cumulative incidences curves for each risk and group.
}
\examples{
fitC <- fitCuminc(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
plotCuminc(ci = fitC, risk = "event", group = "gender", target = 1200)
}
