% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativeIncidences.R
\name{plotCuminc}
\alias{plotCuminc}
\title{making cuminc data for plotting and testing}
\usage{
plotCuminc(time, risk, group, data, target, rho = 0, cens = 0)
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{target}{conf intervals point}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}
}
\description{
list with cumulative incidences and a test of differences between them
}
