% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCuminc.R
\name{plotCuminc}
\alias{plotCuminc}
\title{Cumulative incidences curves}
\usage{
plotCuminc(ci, risk, group, target = NULL, ggtheme = theme_minimal(),
  titleCuminc = "Cumulative incidence function", xtitle = "Time",
  ytitleCuminc = "Cumulative incidences", legendtitle = "Group")
}
\arguments{
\item{ci}{a result of function fitCuminc.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{name of a column indicating group variable, can be numeric or factor/character.}

\item{target}{point in time, in which the confidence bounds should be plotted (default NULL, no confidence bounds plotted).}

\item{ggtheme}{ggtheme to be used (default: theme_minimal()).}

\item{titleCuminc}{a title of a plot (default: "Cumulative incidence function").}

\item{xtitle}{a title of x axis (default: "Time").}

\item{ytitleCuminc}{a title of y axis (default: "Cumulative incidences")}

\item{legendtitle}{a title of a legend (default: "Group").}
}
\value{
a ggplot containing n graphs, where n is number of risks. Each graph represents cumulative incidence curves for given risk in each group.
}
\description{
The function plots cumulative incidences curves for each risk and group.
}
\examples{
fitC <- fitCuminc(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
plotCuminc(ci = fitC, risk = "event", group = "gender", target = 1200)
}
