% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurvival.R
\name{plotSurvival}
\alias{plotSurvival}
\title{Survival curves}
\usage{
plotSurvival(fit, target)
}
\arguments{
\item{fit}{a result of fitSurvival function.}

\item{target}{point in time, in which the confidence bounds should be plotted.}
}
\value{
a ggplot containing n graphs, where n is number of risks. Each graphs represents survival curves for given risk. One curve corresponds to one group.
}
\description{
The function plots survival curves for each risk and group.
}
\examples{
fitS <- fitSurvival(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive", type = "kaplan-meier", conf.int = 0.95, conf.type = "log")
plotSurvival(fit = fitS, target = 1200)
}
