% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCox.R
\name{simpleCox}
\alias{simpleCox}
\alias{fitCox}
\title{Cox model for each type of event separately.}
\usage{
fitCox(time, risk, group, data, cens = 0)
}
\arguments{
\item{time}{name of a column indicating time of an event or follow-up, must be numeric.}

\item{risk}{name of a column indicating type of event, can be numeric or factor/character.}

\item{group}{nam of a column indicating grouping variable, can be numeric or factor/character.}

\item{data}{data.frame, data.table or matrix containing time, risk and group columns.}

\item{cens}{value of 'risk' indicating censored observation (default 0).}
}
\value{
a list of length n, where n is number of risks. Each element of a list is a result of summary.coxph function from package survival, where there is only one type of event possible (other are treating as consored).
}
\description{
The function fits Cox models for each type of event
treating other events as censoring.
}
\examples{
fitCox(time = "time", risk = "event", group = "gender", data = LUAD, cens = "alive")
}
