% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCox.R
\name{testCox}
\alias{testCox}
\title{Testing differences between groups in Cox models.}
\usage{
testCox(fitCox)
}
\arguments{
\item{fitCox}{a result of function fitCox.}
}
\value{
a data.frame with p-values of 3 tests for each risk.
}
\description{
The function provides 3 tests for comparing models estimated from fitCox function. The tests are: Likelihood Ratio Test, Wald Test and Logrank Test.
}
\examples{
fitC <- fitCox(time = LUAD$time, risk = LUAD$event, group = LUAD$gender, cens = "alive")
testCox(fitC)
}
\seealso{
\code{\link[cr17]{fitCox}}
}
