% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit&testCuminc.R
\name{testCuminc}
\alias{testCuminc}
\title{Log rank test for competing risks}
\usage{
testCuminc(ci)
}
\arguments{
\item{time}{time must be numeric}

\item{risk}{can be numeric or factor/character}

\item{group}{can be numeric or factor/character}

\item{data}{can be data frame or matrix}

\item{conf.int}{level of two sided conf int}

\item{conf.type}{"none", "plain", "log" (default), "log-log}
}
\description{
Log rank test for competing risks
}
