% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestCuminc.R
\name{testCuminc}
\alias{testCuminc}
\title{K-sample Test for Competing Risks}
\usage{
testCuminc(ci)
}
\arguments{
\item{ci}{a result of fitCumin function.}
}
\value{
data.frame containing p-values of K-sample test for each risk.
}
\description{
tests differences in cumulative incidences function between groups using K-sample test.
}
\examples{
fitC <- fitCuminc(time = LUAD$time, risk = LUAD$event, group = LUAD$gender, cens = "alive")
testCuminc(fitC)
}
\seealso{
\code{\link[cr17]{fitCuminc}}
}
